# docs/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProvisioningBrokerDeadLetterJobResponse": {
        "properties": {
          "attempts": {
            "default": 0,
            "title": "Attempts",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          },
          "job_type": {
            "title": "Job Type",
            "type": "string"
          },
          "max_attempts": {
            "default": 3,
            "title": "Max Attempts",
            "type": "integer"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "tenant_id",
          "job_type",
          "status",
          "recorded_at"
        ],
        "title": "ProvisioningBrokerDeadLetterJobResponse",
        "type": "object"
      },
      "ProvisioningBrokerDeadLetterResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProvisioningBrokerDeadLetterJobResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "total_jobs": {
            "title": "Total Jobs",
            "type": "integer"
          }
        },
        "required": [
          "success",
          "message",
          "total_jobs",
          "data"
        ],
        "title": "ProvisioningBrokerDeadLetterResponse",
        "type": "object"
      },
      "ProvisioningBrokerRequeueRequest": {
        "properties": {
          "delay_seconds": {
            "default": 0,
            "title": "Delay Seconds",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_contains": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Contains"
          },
          "job_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Type"
          },
          "limit": {
            "default": 50,
            "title": "Limit",
            "type": "integer"
          },
          "reset_attempts": {
            "default": true,
            "title": "Reset Attempts",
            "type": "boolean"
          },
          "tenant_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Slug"
          }
        },
        "title": "ProvisioningBrokerRequeueRequest",
        "type": "object"
      },
      "ProvisioningBrokerRequeueResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProvisioningJobResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "total_jobs": {
            "title": "Total Jobs",
            "type": "integer"
          }
        },
        "required": [
          "success",
          "message",
          "total_jobs",
          "data"
        ],
        "title": "ProvisioningBrokerRequeueResponse",
        "type": "object"
      },
...
      "ProvisioningJobResponse": {
        "properties": {
          "attempts": {
            "default": 0,
            "title": "Attempts",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "job_type": {
            "title": "Job Type",
            "type": "string"
          },
          "max_attempts": {
            "default": 3,
            "title": "Max Attempts",
            "type": "integer"
          },
          "next_retry_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Retry At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "job_type",
          "status"
        ],
        "title": "ProvisioningJobResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Platform Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/platform/provisioning-jobs/broker/dlq": {
      "get": {
        "operationId": "provisioning_broker_dead_letter_jobs_platform_provisioning_jobs_broker_dlq_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "job_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Type"
            }
          },
          {
            "in": "query",
            "name": "tenant_slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Slug"
            }
          },
          {
            "in": "query",
            "name": "error_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error Code"
            }
          },
          {
            "in": "query",
            "name": "error_contains",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error Contains"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisioningBrokerDeadLetterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provisioning Broker Dead Letter Jobs",
        "tags": [
          "platform-provisioning"
        ]
      }
    },
    "/platform/provisioning-jobs/broker/dlq/requeue": {
      "post": {
        "operationId": "requeue_provisioning_broker_dead_letter_jobs_platform_provisioning_jobs_broker_dlq_requeue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisioningBrokerRequeueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisioningBrokerRequeueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Requeue Provisioning Broker Dead Letter Jobs",
        "tags": [
          "platform-provisioning"
        ]
      }
    },
...
    "/platform/provisioning-jobs/{job_id}/requeue": {
      "post": {
        "operationId": "requeue_provisioning_job_platform_provisioning_jobs__job_id__requeue_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "reset_attempts",
            "required": false,
            "schema": {
              "default": true,
              "title": "Reset Attempts",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "delay_seconds",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Delay Seconds",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisioningJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Requeue Provisioning Job",
        "tags": [
          "platform-provisioning"
        ]
      }
    },
    "/platform/provisioning-jobs/{job_id}/run": {
      "post": {
        "operationId": "run_provisioning_job_platform_provisioning_jobs__job_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisioningJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Provisioning Job",
        "tags": [
          "platform-provisioning"
        ]
      }
    },
...
  }
}